Move `AggregateField` to the `aggregate` package
... so that it's in the same package with `AggregateStorage`.

Also make some classes `final`.
